feat: add temporary block subject identities manager and commands
