+++ apps/authenticator/lib/sign_in/schemas/user_attempt.ex
@@ -47,4 +47,12 @@ defmodule Authenticator.SignIn.Schemas.UserAttempt do
   defp custom_query(query, {:ids, ids}), do: where(query, [c], c.id in ^ids)
   defp custom_query(query, {:created_after, date}), do: where(query, [c], c.inserted_at > ^date)
   defp custom_query(query, {:created_before, date}), do: where(query, [c], c.inserted_at < ^date)
+
+  defp custom_query(query, {:temporarilly_blocked, {max_attempts, date}}) do
+    query
+    |> where([c], c.inserted_at > ^date)
+    |> group_by([c], c.username)
+    |> having([c], count(c.username) > ^max_attempts)
+    |> select([c], c.username)
+  end
 end
