--- apps/resource_manager/lib/identity/client_applications.ex
@@ -1,12 +1,0 @@
-defmodule ResourceManager.Identity.ClientApplications do
-  @moduledoc """
-  Client application are subject identities that are not impersonated by a person.
-
-  A client application is allowed to do certain actions by authenticating in the authentication
-  provider with success.
-
-  What a client application can do is defined by it's scopes.
-  """
-
-  use ResourceManager.Domain, schema_model: ResourceManager.Identity.Schemas.ClientApplication
-end
