+++ apps/resource_manager/lib/permissions/commands/consent_scope.ex
@@ -6,7 +6,7 @@ defmodule ResourceManager.Permissions.Commands.ConsentScope do
   require Logger
 
   alias Ecto.Multi
-  alias ResourceManager.Identity.Schemas.{ClientApplication, User}
+  alias ResourceManager.Identities.Schemas.{ClientApplication, User}
   alias ResourceManager.Permissions.Schemas.{ClientApplicationScope, UserScope}
   alias ResourceManager.Repo
