+++ apps/resource_manager/test/support/mocks.ex
@@ -2,7 +2,10 @@ for module <- [
       # Credential ports mocks
       ResourceManager.Credentials.Ports.GenerateHash,
       ResourceManager.Credentials.Ports.VerifyHash,
-      ResourceManager.Credentials.Ports.FakeVerifyHash
+      ResourceManager.Credentials.Ports.FakeVerifyHash,
+
+      # Identity ports mocks
+      ResourceManager.Identities.Ports.GetTemporarillyBlocked
     ] do
   Mox.defmock(:"#{module}Mock", for: module)
 end
